requirements.txt
apache-airflow
psycopg2-binary
requests
streamlit
pandas
sqlalchemy